Allow passthrough rad blocks
Previously, all rad/request blocks required some form of JSON parsing.
However, we should support the usecase of performing a simple query and
print of the raw output, particularly for the case where the response is
not JSON, and thus we don't want to extract JSON fields.

With this commit, we allow omitting the fields RadStmt, which was
previously always required. By omitting fields, the RSL developer is
implicitly saying 'this is a passthrough'. This is intuitive if you
think of rad blocks as being 'request, print raw' as the default, and
the 'fields' stmt is a modifier which allows trying to parse it as JSON
and extracting fields.