Implement rad block 'truncate' statement
  rad url:
      fields <fields>
      truncate <field> <max length>

e.g.

  rad url:
      fields id, name
      truncate name 5

Allows explicitly limiting the size of columns. In the above example, we
limit the 'name' column to 5 characters. Anything longer will only
display (in this example) 4 characters and then ellipses (…).

With our existing 'terminal table fitting' algorithm, we may shorten
truncated columns even further still, if needed, but this at least lets
you explicitly shorten a column if it's known to be long.