Implement rad block 'color' field modifier
Syntax example:

  rad url:
    fields name, city
    name:
      color "red" "o[a-z]"

specifically, the color stmt is

  color EXPR EXPR

where the first expr is the color, allowing for some pre-defined colors.
The reason it's an expression rather than another keyword is so that
your can allow using variables in its place (opening devs up to e.g.
conditional colors or choosing as an arg).

Coloring a column this way will color any segment in a given cell's text
which matches the regex.

Implementation involves some changes over on amterp/go-tbl to allow for
this.

We include a test that the coloring works. I'd been avoiding it so far
but there's no way around it here if we want useful testing for this
feature, which we really do, given the relative complexity of it
(accounting for the complexity in go-tbl). It's a really horrid test
though in terms of readability/writeability, can't think of a better
approach atm. See 'TestRadColorStmts'.