Support multiple field color modifiers on one column
Initial implementation only supported a single color statement.
Specifically, if you specified

  color "red" "New"
  color "blue" "York"

Then 'New' is left blank and only 'York' is colored (blue) in the string
"New York". Of course, we'd want 'New' to be red, and 'York' blue.

We implement this, making later statements take precedence in case of
overlaps. For example,

  color "red" "New"
  color "blue" "w York"

will color 'Ne' in red and 'w York' in blue.

The bulk of the algo and complexity is in the v0.7.0 go-tbl commit. Due
to this complexity, we make sure to add a few more tests (awful as they
are to write for these colored tables).