Do not wrap non-array json fields in arrays
Let's say we have this json

{
  "len": 2,
  "results": [
    {
      "name": "Alice",
      "age": 30
    },
    {
      "name": "Bob",
      "age": 40
    }
  ]
}

Prior to this commit, we'd put *any* capture into an array. i.e. these
two fields:

  length = json.len
  ages = json.results[].age

would respectively be '[2]' and '[30, 40]'. However, it's evident from
looking at length's definition that its capture will not result in an
array (unless 'len' itself is an array, but in that case we wrap *it* in
array, so we'd get a nested array).

This behavior is suboptimal and perhaps suprising. Really, 'length'
should simply be equal to an int 2 after capture.

Let's implement this. Any definition which has a wildcard * or [] in its
definition will get captured into an array. Otherwise, it will not.