Allow capturing json nodes
If we've got a JSON response like this:

{
  "len": 2,
  "results": {
    "Alice": {
      "age": 30,
      "hometown": "New York"
    },
    "Bob": {
      "age": 40,
      "hometown": "Los Angeles"
    }
  }
}

then with this change, we can now capture a json node as a string, which
was not possible before:

  alice = json.results.Alice
  print(alice)

this will print a string '{"age":30,"hometown":"New York"}'

If we later improve Rad's capabilities around passing our *own* JSON to
rad blocks, that means we could probably capture json nodes and further
process it later in the script. We can also simply print the JSON
as-is, if useful/desired.