Implement pprint (pretty print) function
This is usable with any arg types, but is particularly intended for json
strings, probably captured via rad blocks. We use
github.com/nwidger/jsoncolor, which seems like a solid library for this
(gron uses it), and it also serves us quite well because it uses
fatih/color under the hood, which is also what we use in amterp/go-tbl,
so will respect the --NO-COLOR flag almost for free.

My implementation of pprint is complicated a bit by needing to handle
'mixed' json data, meaning we might have a value which is modeled as an
array, but then each item in the array is a json string, but we need to
pretty print the whole thing as one. Definitely a bit rushed and not
properly thought through, should revisit, especially if it proves buggy
(though a bit of unit testing makes it look okay).

pprint only taking one arg, unlike print, appears consistent with how
pprint is implemented in Python as well, as a sanity check. Passing
multiple args is probably not that likely to be useful.