Remove RuntimeLiterals struct
Never really leveraged this construct. There was *one* place where I
checked the 'type' field on it, which was the whole point of its
existence. Everywhere else, we're not dealing with a RuntimeLiteral, but
just handling interface{}, which has not been as bad as I thought.

Let's get rid of this and simplify a bit.