Remove non-mixed array types
The value of typed arrays e.g. int[], float[], etc, stops beyond the
args block, I think. Right now, we also allow explicitly typing arrays
as declarations in the RSL script e.g.

a int[] = [1, 2]

but to be honest, for a scripting language like RSL, I think this is
just a needless complication of the syntax and its implementation.
There's probably very little reason to allow this, and by removing that
sort of thing in this commit, forcing all arrays to be backed by simply
a mixed array i.e. []interface{}, we can *greatly* simplify the
implementation of any value-handling code, cutting the number of switch
cases in half, and also getting rid of some of the really painful
handling in interpreter_long. This reduction can be seen in the deleted
line count (+130 -740 as of writing, though tbf, that's also from
deleting some now-irrelevant tests).

One slight complication is that, since we still want to allow these
typed arrays in the args block, we need to split the enum which
represents Rsl types into two: one for args, and one for 'regular'
running, the latter not containing any of the typed arrays. Similarly, I
anticipate implementing maps in a future commit, and I don't know that
we'll definitely support that in an args block, so only the non-Args
version of the Rsl types enum will contain 'map'.