Implement maps
Originally I was not going to add this cause I felt it'd be unnecessary
for the scope of RSL, but I've come to see that, if we want JSON
handling to be smooth and painless in RSL, then we need a map type, so
that we have RSL types 100% matching what is possible with JSON.

My vision for maps are fairly standard, inspired by languages like
Python. Can define a map like this:

 a = { "alice": 35, "bob": "bar", "charlie": [1, "hi"] }

And look up values like this

  myValue = a["alice"]

Map declarations are made up of expressions, so something like this
works as you'd expect:

  foo = "bar"
  t = true
  f = false
  a = { "alice": 30 + 5, "bob": foo, upper("charlie"): [1, t or f] }

Right now we only allow string values, since that is what JSON does too,
but this might turn out to be a needless limitation on RSL maps
themselves.

We merge 'array' and 'map' access into one interpreted instruction,
since they're so similar.

While here, I noticed that and/or statements are actually broken? Not
sure if this was always the case, but it's a clear gap in test coverage
too, so we fix up the prod code and add coverage.

The map implementation is an ordered map. I avoided bringing in another
dependency to provide a properly implemented e.g. linked hashmap, since
I'd rather avoid adding more dependencies for something that's so
trivial to add like I've done in this commit.

map_test.go contains a bunch of todos that I still need to do.
Immediately are key-based assignments, and utilities like contains,
keyset, iteration.

Also add some explicit len tests since that was missing.