Implement collection (array/map) entry assignment
This includes compound operators. So the following now works as you'd
expect:

  a = [100, 200, 300, 400]
  a[0] += 20
  a[1] -= 20
  a[2] *= 2
  a[3] /= 2
  // final result: [120, 180, 600, 200]

and for maps:

  a = { "alice": 100, "bob": 200, "charlie": 300, "dave": 400 }
  a["alice"] += 20
  a["bob"] -= 20
  a["charlie"] *= 2
  a["dave"] /= 2
  // final result: { alice: 120, bob: 180, charlie: 600, dave: 200 }

Once again, we implement both array and map operations with a single
instruction type, as they're so similar. We also refactor some existing
operation execution code so we can re-use it for this case here.

I also include a WIP 'principles.md' doc for RSL, just so I can begin to
crystalize and note down some of the guiding principles as I've been
working on RSL.