Implement 'del' (delete) statements
Now that we have maps, the ability to remove from collections becomes
more important. I don't have a particular use case atm but I think it'll
inevitably be seen as a 'basic' utility that'll be needed at *some*
point by *someone*. The delete statement also works on whole variables,
deleting them from the environment.

Some examples of using the 'del' statement:

  del a            // delete variable 'a'
  del b, c         // delete both variables 'b' and 'c'
  del d[1]         // delete the second element of array d, shifting elements
                   // on the right to the left
  del e["key"]     // delete the entry with key 'key' in map 'e'
  del f[0]['key']  // delete the entry with key 'key' from a map which
                   // is the first element of array 'f'

The implementation gets a little complex in order to support the latter
case. Since we've make collections stored by value (not reference) in
our environment, if we were to delete an entry from a nested collection,
then this will not trivially get remembered -- we need to re-add the
modified variable into our environment, replacing the old non-modified
value. Part of the solution here is a recursive modification of a given
value (see interpreter.go#executeDelete).

In case it's not obvious, this 'del' syntax and rules are heavily
inspired by Python, as is becoming a tradition :)