Implement 'in' and 'not in' operators
Once again, inspired by Python and works very similarly to it. Can now
do the following as bool-returning expressions:

  "li" in "alice"
  "li" not in "alice"
  myVar in myArray
  myVar in myMap

'myVar in myArray' will do a linear search through the array values.
'myVar in myMap' will attempt to do a direct lookup in the underlying
map and return true if a value was returned i.e. 'myVar' is a key in the
map.

This change involves adding a new line into the EBNF called
'membership'. It slots between equality and comparison.

A bit of refactoring in interpreter_long.go: I use switch type coercion
throughout it now, since I learned that was a Go feature :)

I also finish implementing operations on bool, kinda surprised I haven't
added it sooner, guess there's been no need.

With 'not' now being a keyword used to negate the 'in' operation, we're
a little inconsistent by having chosen ! as the negation operator
boolean expressions. I think I just need to change that to a 'not'
keyword as well, and let the bang go.

My implementation of 'not in' might be a liiiittle inconsistent - it's
the only place in our parser creates new tokens, which so far has 100%
been in the lexer. I'm not sure if that'll be a problem down the road,
since the created tokens are derived from previous tokens, so no
information gets dropped/created out of nowhere and for no reason. I do
this so that I can merge the 'not' and 'in' tokens into one, making the
BinaryExpr emission and implementation easier.