Capture json dicts from rad requests into RslMaps
The previous few commits have implemented a bunch of map-related
functionality in RSL, but so far, the only way to get a map into your
script is to instantiate one yourself with your own data.

The point of adding maps is not that, but instead to get the JSON
captured via rad blocks to be capturable via RslMaps, to make dealing
with json dicts much nicer.

We that that in this commit!

This is also kinda nice cause it let's us clean up a couple of
'hacks'/workarounds I've had to implement due to missing a map type,
such as treating json dicts as strings, marshaling and unmarshaling for
prettify, etc.

I'm noticing that the missing space after the colon and before the value
in jsoncolor's MarshalIndentWithFormatter function is staying around,
and I can't seem to get rid of it. It doesn't happen with Go's built-in
json formatter, so must be a bug in the library. The last commit to that
repo was almost a year ago, unfortunately, so I'm hoping I don't need to
fix it *in* the library, otherwise I may need to fork it. We'll see.