Implement string/array slicing
Inspired by Python, so syntax like this works:

  a[1:3]
  a[:-1]
  a[-99:99]
  a[3:]

We don't support stepwise slices though, like Python does e.g.
a[1:5:10], I just don't think it'd be that useful. Wait and see.

Also fix a small bug in the 'primary' parsing which meant we would stop
as soon as we found a literal, rather than continuing to parse to see if
there were more to it e.g. a slice or index.