Implement ternary expressions
I've gone for the C-like

 <condition> ? <true> : <false>

syntax. Potentially contentious, given we've been going for a
Python-inspired syntax in many areas, but I just personally prefer
thinking about the 'condition' first, and then the true/false branches.
The syntax is also slightly less verbose than the Python

 <true> if <condition> else <false>

syntax. The "? :" style of ternary could be seen as going against the
'readability' / 'no cryptic symbols' principles of RSL, but I think
there's a very safe counter-case to be made that we can rely on devs to
know what "? :" are, since it's so common in C-rooted languages.