Replace '!' negation with 'not' negation
I initially chose '!' because it was terser while still being a
completely standard thing that devs would know from other languages.
However, I've since adopted the Python-style 'not in' syntax for
checking if an item exists in a collection or not. Having that but then
not allowing 'not' as the negation, and expecting '!' instead, is
inconsistent and will probably surprise users. We also don't want to
allow both (I think), since it's better just to have one way to these
sorts of things.

Thus, we replace '!' with 'not'. To demonstrate:

  !true
  !!!true

becomes

 not true
 not not not true