Implement inline expressions in string interpolations
Before, you could only use variables in string interpolations

  greeting = "Hi, {name}"

now, you can do full expressions between the brackets, python-style

  greening = "Hi, {upper(name)}"

Implementing this is unfortunately quite complicated and has some
downsides in terms of complicating how string literals get emitted.

Prior, we'd emit the whole string as a single token string literal

 age = 2 + 2
 "Age: {age}"

Now, we emit several tokens. If we were to inline above:

 "Age: {2 + 2}!"

this gets emitted as these tokens:

 "Age: "  < string literal
 2        < number literal
 +        < plus literal
 2        < number literal
 "!"      < string literal

String literals now contain information on them to indicate if they're
followed by an inline expression, so that the parser can tie that whole
group of tokens into one visitable literal which can evaluate any inline
expressions and produce one final string in the interpreter.

Because we also want to reference it in some places, and for errors,
each string literal contains a 'FullStringLiteral', which is really the
full string as it's been parsed up until that string literal. That means
that the *last* string literal in a given series of literals is the one
that contains the *real* FullStringLiteral, i.e. in the above case:

 "Age: " contains just 'Age: ' as the 'FullStringLiteral'
 "!" correctly contains "Age: {2 + 2}!" as the 'FullStringLiteral'

That's a bit of an ugly way to do this... But I can't think of a nice
way to do right now, and don't want to block progress.

We also put things in place to be able to do formatting as well, e.g.
{pi:.2f}.