Implement instring expression formatting
New term 'instring expression': an expression expected in a string
between {} braces, for evaluation.

This commit enables formatting for these scenarios, inspired by Python.
Some examples:

 {var:16}    // pad with spaces up to 16 chars
 {var:16.2}  // ... and with 2 decimal places (expect number)
 {var:<16}   // left-align the padded value
 {var:.2f}   // print 2 decimal places, no padding

The part prior to : can be any expression:

 {1 + 2.14:.1} // prints '3.1'

We'd like to handle 0-padding, but this is not currently supported.
Also, because we delegate the formatting to Go's fmt.Sprintf, then
unless a precision is specified, it will 6 decimal places, even if the
latter places are 0s and arguably unnecessary. This too, I think we'd
like to get rid of, but is complicated a bit by correctly handling cases
that also have padding.

The implementation isn't the *cleanest*, but I think it does the job
well enough at the moment.