Allow '_' variable names
Noticed the lexer was only allowing variable names starting with a
letter. Maybe I was being conservative, but at this point, I think it's
something we can allow without regretting it - having a _ var name is a
common convention for signaling you don't really care about the variable
name so just toss it.

A minor ordering thing for for i-v loop implementation: previously, we
were setting the 'v' *before* the 'i' in the environment, so if you do
for _, _ and then tried to reference '_', it'd actually refer to the
index despite it coming *first*, not the value.

Separately, maybe we should disallow referencing _ vars, but I'll leave
it for now (just adds complexity to RSL's implementation).