Support ` backticks as string delimiters
We now support 3 ways to make strings:

1. "double quotes"
2. 'single quotes'
3. `backticks`

The purpose of having multiple is so that, if your content itself needs
to contain any of the delimiters, you have options for how to wrap the
whole thing in a way which won't require you to escape any of those
characters.

Backticks will be particularly useful for shell commands, as I intend to
implement, since they're not commonly used there.