Implement (roughly) shell commands
This is quite tricky to get right, and there's still a bit of work
remaining, so more to come, but this commit is the core of how it will
work.

Basically, you can invoke any string as a shell command by putting a $
in front of it:

 $"ls -l"

the recommended/idiomatic way to do this will probably be to use
backticks, since they will have fewer conflicts with shell cmd contents:

 $`ls -l`

The string is an expression so you can do

 cmd = `ls- l`
 $cmd

I've left a bunch of todos up at the top of interpreter_shell.go, some
of which will be next in the following commits.

I think the final syntax will be:

 err_code = $`ls -l`
 failure:
     print("how'd this fail?!")
     exit(err_code)

where line 1 can also be

 err_code, stdout, stderr = $`ls -l`

The difference is that, in the first, we just print stdout and stderr
from the command immediately. In the second, we'll *instead* capture it
into the vars, and the command becomes silent. It's like redirecting.
You can have stdout and leave stderr if you want, and stderr will be the
only one printed.

You can also do this if you really don't want the failure block (which
will be required otherwise):

 err_code = unsafe $`ls -l`

Anyway, all these things are not how it is yet in this commit. Hopefully
soon!

I also included a VERSION_LOG.md, to try and remember/list the major
items included in each version.