Improve shell command execution / output capture
With this commit, I think the shell commands feature is pretty good.
I'll definitely wanna follow up with tests and a little bit more polish,
but this is good and has already allowed me to turn some
bash-by-necessity scripts into full Rad scripts, as I can execute shell
commands now in those Rad scripts.

One important change is we reorder the unpacking order:

  error_code, stdout, stderr = unsafe $`ls -l`

Before, the error code was last, now it's first. Part of the reason is
that capturing stdout and stderr having increasingly large impacts on
the command being run and its output. As soon as you start capturing
stdout or stderr, it will no longer get printed to the terminal. This is
a compromise we've had to make because some commands just do not play
nice in the scenario where they don't have direct output to the
terminal. Trying to support this just leads to wacky output and I don't
wanna try to tackle that, at least at the moment. So, by having the
error code *first*, you can capture that and use it later, without
impact stdout or stderr.

As I've mentioned there, another important difference with this commit
is that we only capture stdout/stderr if their assignment identifiers
are defined.

I make a couple of small changes to the Makefile.

- build no longer inherits from format. As-is, build was equivalent to
  'all' because of that inheritence, and sometimes you just wanna
  rebuild without rerunning the code gen, because you know that won't
  change.
- build now outputs to a git ignore'd bin directory. the resulting
  binary is also called 'radd'. the purpose is so that I can put the bin
  folder on my path and thus use 'radd' on my computer, as a dev build.