Allow negative default int/float args
Previously, this would fail to get parsed properly:

  args:
    myArg int = -10

That's because the '-' and '10' are separate tokens. Everywhere else in
RSL, when we parse numbers and that sort of thing, we do it via the
'expr' AST. This includes a unary step which will check for minuses and
handle that sort of thing.

However, for arg defaults, we intentionally read only literal tokens.
Unfortunately this misses minuses and the parser doesn't know what to do
with it.

Solving this isn't super nice. Making it a true 'expr' *would* be
nicest, but we can't trivially do that while statically checking that
the default matches the type of the arg. It's possible, but not
something I wanna tackle right now.

The fix here is to, in the 'literal' parser, prior to reading
ints/floats, check for minuses, and consume all of them until there
are no more. If there's an odd number of minuses, the number will be
negative, which we pass down to the resulting literal for use in the
interpreter.

I also have to slightly alter the file header lexing so that triple
negative doesn't just throw there.

While here, I also convert the push.sh script to a full RSL script
push.rad. It's not the *main* use case for rad but it's arguably nicer
and crucially it allows for more dog-fooding of Rad for things in this
repo.