Change shell cmds: add critical commands, fail/recover blocks
I don't I had fully thought through the previous approach, what I really
wanted was a way to run commands and simply have the script exit at that
point if the command failed. That's not what the previous approach did.

With this commit, we introduce the idea of a 'critical' shell command.
The syntax is:

  $!`ls -l`

i.e. an exclamation aka 'bang' after the dollar sign. Critical shell
commands, if they fail, will print the error and exit the RSL script
then and there.

I was tossing up a little what the syntax should be. The bang is
*slightly* antithetical to our philosophy so far, which is for scripts
to be self-explanatory and not rely on random symbology. An alternative
was 'must' or 'require', for example:

  must $`ls -l`

or

  require $`ls -l`

These are more self-explanatory, but I don't think they're *so much
more* self explanatory that they justify needing to be written every
time an RSL dev wants to execute a critical shell command. What I
anticipate will be the case, is that a large % of executed shell
commands will be critical. And so, I think 1) people will get used to
the syntax quickly and 2) the terse-ness and benefits to writeability
(and readability, once familiar) outweigh the cons for the bang.

It's also not entirely without precedent. It's inspired by Kotlin (and
TypeScript/JS)'s non-null assertion operator, so it hopefully will be
familiar to devs familiar with those.

The other piece to this change is the removal of the 'failure' block,
and its replacement with two new ones: fail and recover. They can be
used like this:

  err = $`ls -l`
  fail:
    print("How'd this fail?!")

  print("Alice")

or

  err = $`ls -l`
  recover:
    print("How'd this fail?!")

  print("Alice")

These blocks only get run if the attached shell command fails. For the
'fail' block, it will get run, and the RSL script will exit after.

For the 'recover' block, that block will get run, and the RSL script
will continue. In the above examples, "Alice" only gets printed in the
latter 'recover' example.

In this way, the 'recover' block is a actually like the previous
'failure' block, and the 'fail' block is new. The 'fail' block is a more
concise and preferred way to basically print *something* useful to the
user to give more context if a shell command fails.