Implement 'defer' statements
Inspired by Go, I've found myself wanting these in a couple of places,
so probably makes sense to add to RSL.

Syntax is pretty straight-forward:

  defer print("hi 1")
  defer:
    print("hi 2")
    print("hi 3")

The above shows the single-statement version, and the block version.

Defer statements run in a LIFO order, to try and facilitate correct
ordering. This aligns with how other languages do this.

If you were to run the snippet above, the output would be:

  hi 2
  hi 3
  hi 1

Implementation-wise, we do some somewhat tricky finessing to get correct
behavior, in order to correctly handle cases where a defer statement
itself may fail.

Basically, we wrap the Rad exit function to invoke any registered
deferred statements first. On the first invocation, we note it's the
first invocation, and then begin invoking the deferred statements. If we
then ever make it into the exit block again, it's because we failed one
or a defer statement itself called exit (strange, whatever). We thus
panic so that we can propagate back up to where the deferred statement
execution is happening, and continue on with the rest.

This also makes me think about whether a 'defer on error' syntax would
make sense. Apparently Zig has something, 'errdefer', so it might not be
a crazy idea. Could be useful for rolling back script actions in the
event it fails. Anyway, it's on the potential to-do list.