Fix incorrect DEDENT/NEWLINE token consumption in rad sort parsing
Bit of a specific bug, but I managed to run into earlier.

I can probably manifest in several cases, but for the one I saw, if you
have an if block which contains a rad statement which ends with a sort
statement, that sort statement will erroneously consume a DEDENT token,
which we really should be leaving for the if-block to consume, as a sign
that the block is done. Here's an example where this would go wrong
before:

if true:
  rad url:
    fields name, age, city
    sort age asc
else:
  print("whatever")

The if-else parsing relies on a DEDENT to know the 'if' part is done,
and then resume with the else, but it was prematurely 'eaten' by the rad
sort.

Fix is pretty simple -- make the rad statement terminate when it *peeks*
a newline, and not look at DEDENTs.

A couple of misc cleanups included in this commit.