Add if statements to rad blocks
When you enter a 'rad' block, only certain syntax is allowed. Prior to
this commit, if statements were not allowed, so if you wanted to make a
rad block subject to parameters, you'd need to duplicate most of the
block. For example:

    if do_not_sort:
        rad url:
            fields Name, Age
            Name:
                truncate 20
    else:
        rad url:
            fields Name, Age
            Name:
                truncate 20
            sort asc

This is pretty annoying, and that's an example with just one binary
difference, and a pretty small rad block. After this commit, the above
can be replaced with:

    rad url:
        fields Name, Age
        Name:
            truncate 20
        if !do_not_sort:
            sort asc

Much nicer!

One big note is that I've removed a bit of static validation from the
parser as part of this change, because keeping it correct, given the if
statement, is a bit more work than I would like to do at the moment, and
I think the idea of 'static validation' deserves a concerted push at
some point anyway, and I'll be able to tackle it better with less parts
moving once RSL is a bit more complete.
The validation in question checked that, based on the statements in the
rad block, if a 'fields' statement was required and if only one had been
set. The reason that's a bit onerous to keep here is because the if
statement makes it conditional. It's *hard*, it's just not where I wanna
spend my energy right now.