Implement 'sleep' function, add named arguments
We implement a 'sleep' function as the only way to sleep right now is
via a shell invocation. That's not necessarily that bad but it's not a
great requirement, and by implementing a sleep function ourselves, we
can also add some bells and whistles to it. For example, inspired by Go,
we allow passing a human-friendly string into the sleep function, like
this:

  sleep("5m30s")

or

  sleep("1h 10m")

I also want to (but only implement the 'title' part in this commit)
allow the following:

  sleep(2, title="Pushing...", spinner=true)  // or progress=true

which could be used as a 'warning' that we're about to push, to give
time for users to cancel if they don't want that.

This means finally implementing named arguments, which I've got some
non-that-important use cases for elsewhere too. It works somewhat like
Python, except that you can *only* configure these via this named
approach -- you cannot keep passing positional args and line them up
that way.

This is for three reasons:

1. Less lock-in on API. If we offer it positionally, then we need to
   make API commitments to the ordering of things, which I'd like to
   avoid.
2. Easier read (self-documenting), and arguably easier to write. We'd
   like to avoid having 'required tribal knowledge' here as to the
   ordering of these 'random' arguments, and we should try avoiding
   requiring people to read docs/references to understand what each arg
   as it might be otherwise unintuitive. To demo, it's not obvious what
   the bool here does:

     sleep(2, "Pushing...", true)

   but it's clearer here:

     sleep(2, title="Pushing...", spinner=true)

   This is less of an issue in some of the 'bigger' languages out there
   because you can, with IDEs, often 'click through' these functions and
   see their signature and the ordering/naming of args, but I don't know
   if that's ever something we can do for RSL (I don't know).
3. Easier to implement :) (not really a reason but is convenient)

The implementation includes the ability to mock the sleep function we
use, so that we can thoroughly unit test this without actually making
the unit tests sleep, which would be terrible.