Json Algo: Allow indexing in json path definitions
If you've got data like this:

  [
    {
      "name": "Alice",
      "date": [2024, 12, 13]
    },
    {
      "name": "Bob",
      "date": [2023, 12, 14]
    }
  ]

Then previous to this commit, if you wanted to grab the years and print
them in a table, you'd need to capture the whole date arrays as an array
of arrays, to some manual processing & extraction, and then do a display
block.

With this commit, you can instead do this in your definition:

  YYYY = json[].date[0]

This will capture only the first element in each date array. After this,
YYYY is equal to [2024, 2023].

The implementation is a bit ugly, this feature doesn't fit neatly into
the algo as-is. I think I'll keep tacking features on like this and,
once I've got a better idea for what will be required of the algo, I'll
re-implement it from scratch.