Move json path assignment out of lexer, entirely into parser
Should've probably been implemented this way from the start -- it's
definitely a parser-level concept.

The catalyst for doing this now is that I want to implement the ability
to add array indexing into the json path definition e.g.

  a = json.results[0].name

which would extract the name of the 0th element into a string rather
than *all* the names into a list.

Implementing this will be easier once everything has already been lexed
and I'm dealing with Tokens rather than raw runes.

Nice bonus that this actually being a nice net negative on lines of
code.