Implement rad block 'map' field modifier
If you want a specific format for how a column gets displayed, then
prior to this commit, you'd need to do a 'request', some manual
processing and formatting, and then a display block with the processed
data.

We can offer something a bit more ergonomic, by allowing a 'map'
operation/lambda as part of rad blocks. For example:

  nums = [0.6342, 0.7, 1.63, 0.0923]
  display:
    fields nums
    nums:
      map num -> "{num * 100:6.2}%"

This will display the column like this (formatting each element):

   63.42%
   70.00%
  163.00%
    9.23%

This seems to work pretty well. You can apply the lambda two several
columns if they're the same type, using the existing multi-field mod
syntax:

  display:
    field FirstName, LastName
    FirstName, LastName
      map name -> upper(name)

^ This will apply upper() to both columns.

This syntax actually makes the 'truncate' syntax somewhat redundant.
Rather than doing this:

  display:
    field Name
    Name:
      truncate 20

you can now do

  display:
    field Name
    Name:
      map name -> name[:20]

This is actually pretty compelling...

Anyway, there's a little more work to do with regards to sorting, before
this change is really in a good spot. Sorting right now is done on the
strings *after* these mappings. It's more a problem with how sorting is
implemented though -- it should be done before we stringify columns i.e.
it should be on the original data. Future commit.