Replace rad truncate with truncate func + map keyword
The previous commit implemented a 'map' field modifier. This is quite
flexible and can easily replace our 'truncate' keyword. Instead of:

  rad url:
    fields words
    words:
      truncate 10

we now do

  rad url:
    fields words
    words:
      map x -> truncate(x, 10)

This might seem like a bad trade, and in a way it is, and I might change
my mind and bring back 'truncate', but this is an opportunity to remove
an entire keyword with minimal tradeoff. It also makes things more
consistent in how things *should* be done. Kinda like that "there should
only be one way to do things" philosophy.

One tricky thing I learned is regarding the UTF character '…'. If you
invoke len("…") (on that char), it actually returns the number of
*bytes* in the string, not the number of chars. '…' is actually 3 bytes,
so it returns 3. That's undesirable for these truncation operations, I'm
more interested in how many places on the terminal the string takes up.
For that, we use utf8.RuneCountInString(), which correctly returns 3.

I am probably doing 'len' erroneously in a lot of places, meaning emojis
could break parts of rad. Might try to audit that at some point.