Fix rad 'if stmt' for non-display blocks
I had naively thought testing just the 'display' case would be enough,
but it was not. The colon parsing is a bit different for
display/rad/request blocks, so it meant that the way I had implemented
rad if statements was only valid for display blocks.

This commit fixes that, in part by simplifying how some of the colon
parsing works.