Print shell commands by default before they run
I found myself wanting (without exception) to 'announce' the shell
commands my scripts are running as its runs them. The only way to do
this was via an explicit print statement prior to invoking the shell
command, but we can save script writers from this by doing it out of the
box. Output looks like this (this is from the push.rsl script):

  ⚡️ Running: go build main.go
  ⚡️ Running: go test ./core/testing -count 30
  ok      rad/core/testing    4.643s

This is when just running those two commands, no explicit print
statements.

The lightning bolt seems like a reasonable default. It's cool and
visible but not overused like the 🚀 emoji.

I also rename my scripts to .rsl instead of .rad. Not super important
and might change my mind, but I think it's probably worth tying the
scripts more closely to the idea of the *language* they're written in,
not the tool that runs them.